Add format presets for `/nick`

+++ src/main/kotlin/listener/ChatListener.kt
@@ -35,6 +35,7 @@ class ChatListener(
 
     @Subscribe
     fun joinEvent(event: LoginEvent) {
+        joinMessage(event)
         val unreadCount = chattORE.database.getMessages(event.player.uniqueId).filter { !it.read }.size
         if (unreadCount > 0)
             chattORE.proxy.scheduler.buildTask(chattORE, Runnable {
@@ -47,13 +48,12 @@ class ChatListener(
         if (!chattORE.config[ChattORESpec.clearNicknameOnChange]) return
         val existingName = chattORE.database.uuidToUsernameCache[event.player.uniqueId] ?: return
         if (existingName == event.player.username) return
+        val nickname = chattORE.database.getNickname(event.player.uniqueId);
+        if (nickname?.contains("<username>") ?: false) return
         chattORE.database.removeNickname(event.player.uniqueId)
     }
 
-    @Subscribe
-    fun joinMessage(event: ServerPostConnectEvent) {
-        if (event.player.uniqueId in chattORE.onlinePlayers) return
-        chattORE.onlinePlayers.add(event.player.uniqueId)
+    fun joinMessage(event: LoginEvent) {
         val username = event.player.username
         chattORE.broadcast(
             chattORE.config[ChattORESpec.format.join].render(mapOf(
@@ -70,8 +70,7 @@ class ChatListener(
 
     @Subscribe
     fun leaveMessage(event: DisconnectEvent) {
-        if (event.player.uniqueId !in chattORE.onlinePlayers) return
-        chattORE.onlinePlayers.remove(event.player.uniqueId)
+        if (event.loginStatus != DisconnectEvent.LoginStatus.SUCCESSFUL_LOGIN) return
         val username = event.player.username
         chattORE.broadcast(
             chattORE.config[ChattORESpec.format.leave].render(mapOf(
